moves te progress header to the main activity
