+++ app/src/main/java/com/walkly/walkly/ui/map/MapFragment.kt
@@ -63,17 +63,7 @@ class MapFragment : Fragment(), OnMapReadyCallback, PermissionsListener {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        Player.level.observe(this, Observer {
-            user_level.text = "LEVEL $it"
-        })
-
-        Player.progress.observe(this, Observer {
-            progressBar2.progress = it.toInt()
-        })
-
-
         val btn_bg = join_button.background
-        updateTopBar()
 
         linearLayout = bottom_sheet
         //hide the bottom sheet
@@ -178,7 +168,6 @@ class MapFragment : Fragment(), OnMapReadyCallback, PermissionsListener {
                 })
 
                 BottomSheetBehavior.from(linearLayout).state = BottomSheetBehavior.STATE_COLLAPSED
-                updateTopBar()
             }
             // }
 
@@ -231,44 +220,7 @@ class MapFragment : Fragment(), OnMapReadyCallback, PermissionsListener {
        // Toast.makeText(this, "NOOOOOOOOOOOOOOOO", Toast.LENGTH_LONG).show()
     }
 
-    fun updateTopBar(){
-        Player.stamina.observe(this, Observer {
-            Log.d("stamina from map2", it.toString())
 
-            join_button.isClickable = true
-            join_button.background.alpha = 255
-
-            if(it >= 300){
-                //3 balls
-                stamina1full.visibility = View.VISIBLE
-                stamina2full.visibility = View.VISIBLE
-                stamina3full.visibility = View.VISIBLE
-
-            }else if(it >= 200 ){
-                //2 balls
-                stamina1full.visibility = View.VISIBLE
-                stamina2full.visibility = View.VISIBLE
-                stamina3full.visibility = View.INVISIBLE
-
-            }else if(it >= 100){
-                //1 ball
-                stamina1full.visibility = View.VISIBLE
-                stamina2full.visibility = View.INVISIBLE
-                stamina3full.visibility = View.INVISIBLE
-
-            }else{
-                //no balls
-                stamina1full.visibility = View.INVISIBLE
-                stamina2full.visibility = View.INVISIBLE
-                stamina3full.visibility = View.INVISIBLE
-
-                // player cannot join a battle
-                join_button.isClickable = false
-                join_button.background.alpha = 100
-            }
-
-        })
-    }
 
     override fun onPermissionResult(granted: Boolean) {
         if (granted) {
