+++ app/src/main/res/layout/activity_main.xml
@@ -2,6 +2,7 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    xmlns:mapbox="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#ccc">
@@ -73,8 +74,136 @@
         app:layout_constraintBottom_toTopOf="@+id/btn_map"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/con2"
         app:navGraph="@navigation/mobile_navigation" />
 
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:id="@+id/con2"
+        android:layout_width="match_parent"
+        android:layout_height="80dp"
+        android:background="@drawable/gradient"
+        mapbox:layout_constraintTop_toTopOf="parent">
+
+        <TextView
+            android:id="@+id/user_level"
+            android:layout_width="75dp"
+            android:layout_height="wrap_content"
+            mapbox:layout_constraintStart_toStartOf="parent"
+            mapbox:layout_constraintEnd_toStartOf="@id/progressBar2"
+            mapbox:layout_constraintTop_toTopOf="@id/progressBar2"
+            android:text="LEVEL 1"
+            android:layout_marginLeft="16dp"
+            android:textColor="#FFFFFF"
+            android:textStyle="bold"
+            />
+
+        <ProgressBar
+            android:id="@+id/progressBar2"
+            style="?android:attr/progressBarStyleHorizontal"
+            android:layout_width="258dp"
+            android:layout_height="12dp"
+            android:background="#FFFEFE"
+            android:max="100"
+            android:progress="50"
+            mapbox:layout_constraintBottom_toBottomOf="parent"
+            mapbox:layout_constraintEnd_toEndOf="parent"
+            mapbox:layout_constraintHorizontal_bias="0.379"
+            mapbox:layout_constraintStart_toStartOf="parent"
+            mapbox:layout_constraintTop_toTopOf="parent"
+            mapbox:layout_constraintVertical_bias="0.235"
+            android:backgroundTint="#00FFFEFE"
+            android:progressDrawable="@drawable/levelbar"
+            android:layout_marginLeft="16dp"
+            />
+
+        <ImageView
+            android:id="@+id/stamina1full"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="56dp"
+            android:layout_marginLeft="56dp"
+            android:layout_marginTop="4dp"
+            android:background="@drawable/staminacircle"
+            android:visibility="invisible"
+            mapbox:layout_constraintStart_toStartOf="parent"
+            mapbox:layout_constraintTop_toBottomOf="@+id/progressBar2" />
+
+
+        <ImageView
+            android:id="@+id/stamina2full"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginLeft="6dp"
+            android:layout_marginTop="4dp"
+            android:background="@drawable/staminacircle"
+            android:visibility="invisible"
+            mapbox:layout_constraintStart_toEndOf="@+id/stamina1full"
+            mapbox:layout_constraintTop_toBottomOf="@+id/progressBar2" />
+
+        <ImageView
+            android:id="@+id/stamina3full"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginLeft="6dp"
+            android:layout_marginTop="4dp"
+            android:background="@drawable/staminacircle"
+            android:visibility="invisible"
+            mapbox:layout_constraintStart_toEndOf="@+id/stamina2full"
+            mapbox:layout_constraintTop_toBottomOf="@+id/progressBar2" />
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            app:layout_constraintStart_toStartOf="@id/user_level"
+            app:layout_constraintTop_toBottomOf="@id/progressBar2"
+            android:layout_marginTop="12dp"
+            android:textColor="@color/white"
+            android:textStyle="bold"
+            android:text="Energy"/>
+
+        <ImageView
+            android:id="@+id/stamina1empty"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="56dp"
+            android:layout_marginLeft="56dp"
+            android:layout_marginTop="4dp"
+            android:background="@drawable/staminacircleempty"
+            mapbox:layout_constraintStart_toStartOf="parent"
+            mapbox:layout_constraintTop_toBottomOf="@+id/progressBar2" />
+
+
+        <ImageView
+            android:id="@+id/stamina2empty"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginLeft="6dp"
+            android:layout_marginTop="4dp"
+            android:background="@drawable/staminacircleempty"
+            mapbox:layout_constraintStart_toEndOf="@+id/stamina1empty"
+            mapbox:layout_constraintTop_toBottomOf="@+id/progressBar2" />
+
+        <ImageView
+            android:id="@+id/stamina3empty"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginLeft="6dp"
+            android:layout_marginTop="4dp"
+            android:background="@drawable/staminacircleempty"
+            mapbox:layout_constraintStart_toEndOf="@+id/stamina2empty"
+            mapbox:layout_constraintTop_toBottomOf="@+id/progressBar2" />
+
+        <ImageView
+            android:id="@+id/menu"
+            android:layout_width="48dp"
+            android:layout_height="48dp"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            app:layout_constraintBottom_toBottomOf="parent"
+            android:layout_marginRight="8dp"
+            android:tint="@color/white"
+            android:src="@drawable/baseline_menu_24"/>
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
