+++ app/src/main/res/layout/fragment_map.xml
@@ -24,118 +24,6 @@
         </com.mapbox.mapboxsdk.maps.MapView>
     </androidx.constraintlayout.widget.ConstraintLayout>
 
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/con2"
-        android:layout_width="match_parent"
-        android:layout_height="80dp"
-        android:background="@drawable/gradient"
-        mapbox:layout_constraintTop_toTopOf="parent">
 
-        <TextView
-            android:id="@+id/user_level"
-            android:layout_width="75dp"
-            android:layout_height="wrap_content"
-            mapbox:layout_constraintStart_toStartOf="parent"
-            mapbox:layout_constraintEnd_toStartOf="@id/progressBar2"
-            mapbox:layout_constraintTop_toTopOf="@id/progressBar2"
-            android:text="LEVEL 1"
-            android:layout_marginLeft="16dp"
-            android:textColor="#FFFFFF"
-            android:textStyle="bold"
-            />
-
-        <ProgressBar
-            android:id="@+id/progressBar2"
-            style="?android:attr/progressBarStyleHorizontal"
-            android:layout_width="258dp"
-            android:layout_height="12dp"
-            android:background="#FFFEFE"
-            android:max="100"
-            android:progress="50"
-            mapbox:layout_constraintBottom_toBottomOf="parent"
-            mapbox:layout_constraintEnd_toEndOf="parent"
-            mapbox:layout_constraintHorizontal_bias="0.379"
-            mapbox:layout_constraintStart_toStartOf="parent"
-            mapbox:layout_constraintTop_toTopOf="parent"
-            mapbox:layout_constraintVertical_bias="0.235"
-            android:backgroundTint="#00FFFEFE"
-            android:progressDrawable="@drawable/levelbar"
-            android:layout_marginLeft="16dp"
-            />
-
-        <ImageView
-            android:id="@+id/stamina1full"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="56dp"
-            android:layout_marginLeft="56dp"
-            android:layout_marginTop="4dp"
-            android:background="@drawable/staminacircle"
-            android:visibility="invisible"
-            mapbox:layout_constraintStart_toStartOf="parent"
-            mapbox:layout_constraintTop_toBottomOf="@+id/progressBar2" />
-
-
-        <ImageView
-            android:id="@+id/stamina2full"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="6dp"
-            android:layout_marginTop="4dp"
-            android:background="@drawable/staminacircle"
-            android:visibility="invisible"
-            mapbox:layout_constraintStart_toEndOf="@+id/stamina1full"
-            mapbox:layout_constraintTop_toBottomOf="@+id/progressBar2" />
-
-        <ImageView
-            android:id="@+id/stamina3full"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="6dp"
-            android:layout_marginTop="4dp"
-            android:background="@drawable/staminacircle"
-            android:visibility="invisible"
-            mapbox:layout_constraintStart_toEndOf="@+id/stamina2full"
-            mapbox:layout_constraintTop_toBottomOf="@+id/progressBar2" />
-
-        <ImageView
-            android:id="@+id/stamina1empty"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="56dp"
-            android:layout_marginLeft="56dp"
-            android:layout_marginTop="4dp"
-            android:background="@drawable/staminacircleempty"
-            mapbox:layout_constraintStart_toStartOf="parent"
-            mapbox:layout_constraintTop_toBottomOf="@+id/progressBar2" />
-
-
-        <ImageView
-            android:id="@+id/stamina2empty"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="6dp"
-            android:layout_marginTop="4dp"
-            android:background="@drawable/staminacircleempty"
-            mapbox:layout_constraintStart_toEndOf="@+id/stamina1empty"
-            mapbox:layout_constraintTop_toBottomOf="@+id/progressBar2" />
-
-        <ImageView
-            android:id="@+id/stamina3empty"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="6dp"
-            android:layout_marginTop="4dp"
-            android:background="@drawable/staminacircleempty"
-            mapbox:layout_constraintStart_toEndOf="@+id/stamina2empty"
-            mapbox:layout_constraintTop_toBottomOf="@+id/progressBar2" />
-
-
-
-
-
-
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
     <include layout="@layout/bottom_sheet_layout"></include>
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
